(feat): set package and pytest
