+++ .gitignore
@@ -2,6 +2,7 @@
 __pycache__/
 *.py[cod]
 *$py.class
+.pypirc
 
 # C extensions
 *.so
