+++ pytest.ini
@@ -1,6 +1,3 @@
 [pytest]
-testpaths = tests
-log_cli = 1
-log_cli_level = WARNING
-log_cli_format = %(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)
-log_cli_date_format=%Y-%m-%d %H:%M:%S
+testpaths = src/tests
+addopts = --cov=src/PyReprism --cov-report=html --cov-report=term --cov-fail-under=80
