+++ requirements.txt
@@ -1,2 +1,3 @@
 pytest
 flake8
+pytest-cov
