+++ src/PyReprism/languages/__init__.py
@@ -1,0 +1,10 @@
+from .python import *
+from .java import *
+from .c import *
+from .csharp import *
+from .bash import *
+from .markup import *
+from .perl import *
+from .php import *
+from .ruby import *
+from .rust import *
