+++ src/PyReprism/languages/bash.py
@@ -1,5 +1,5 @@
 import re
-from utils.file_extensions import FileExtension
+from PyReprism.utils import extension
 
 class Bash:
     def __init__():
@@ -11,7 +11,11 @@ class Bash:
         pattern = re.compile(r'(?P<comment>#.*?$)|(?P<noncomment>\'(\\.|[^\\\'])*\'|"(\\.|[^\\"])*"|.[^#\'"]*)', re.DOTALL | re.MULTILINE)
 
         return pattern
+    
+    @staticmethod
+    def keywords() -> str:
+        pass
         
     @staticmethod
     def file_extension() -> str:
-        return FileExtension.bash
+        return extension.bash
