+++ src/PyReprism/languages/csharp.py
@@ -1,0 +1,17 @@
+import re
+from PyReprism.utils import extension
+class CSharp:
+    def __init__():
+        pass
+
+    @staticmethod
+    def comment():
+        pass
+
+    @staticmethod
+    def keywords():
+        pass
+    
+    @staticmethod
+    def file_extension():
+        return extension.ruby
