+++ src/PyReprism/languages/markup.py
@@ -1,5 +1,5 @@
 import re
-from PyReprism.utils.extensions import FileExtension
+from PyReprism.utils import extension
 
 class MarkUp:
     def __init__():
@@ -10,9 +10,12 @@ class MarkUp:
         
         return re.compile(r'(?P<multilinecomment><!--.*?-->)|(?P<noncomment>\'(\\.|[^\\\'])*\'|"(\\.|[^\\"])*"|.[^/\'"]*)', re.DOTALL | re.MULTILINE)
         
+    @staticmethod
+    def keywords() -> list:
+        pass
     
     @staticmethod
     def file_extension():
-        return FileExtension.markup
+        return extension.markup
