+++ src/PyReprism/languages/rust.py
@@ -1,0 +1,17 @@
+import re
+from PyReprism.utils import extension
+class Rust:
+    def __init__():
+        pass
+
+    @staticmethod
+    def comment():
+        pass
+
+    @staticmethod
+    def keywords() -> list:
+        pass
+        
+    @staticmethod
+    def file_extension():
+        return extension.rust
