--- src/PyReprism/utils/extensions.py
@@ -1,147 +1,0 @@
-class FileExtension:
-    abap: str = "xxxxx"
-    actionscript: str = "xxxxx"
-    ada: str = "xxxxx"
-    apacheconf: str = "xxxxx"
-    apl: str = "xxxxx"
-    applescript: str = "xxxxx"
-    arduino: str = "xxxxx"
-    arff: str = "xxxxx"
-    asciidoc: str = "xxxxx"
-    asm6502: str = "xxxxx"
-    aspnet: str = ".asp"
-    autohotkey: str = "xxxxx"
-    autoit: str = "xxxxx"
-    bash: str = ".sh"
-    basic: str = "xxxxx"
-    batch: str = "xxxxx"
-    bison: str = "xxxxx"
-    brainfuck: str = "xxxxx"
-    bro: str = "xxxxx"
-    c: str = ".c"
-    clike: str = "xxxxx"
-    clojure: str = "xxxxx"
-    coffeescript: str = "xxxxx"
-    cpp: str = ".cpp"
-    crystal: str = "xxxxx"
-    csharp: str = ".csharp"
-    csp: str = "xxxxx"
-    css_extras: str = "xxxxx"
-    css: str = ".css"
-    d: str = "xxxxx"
-    dart: str = ".dart"
-    diff: str = ".diff"
-    django: str = "xxxxx"
-    docker: str = "xxxxx"
-    eiffel: str = "xxxxx"
-    elixir: str = "xxxxx"
-    erb: str = "xxxxx"
-    erlang: str = "xxxxx"
-    flow: str = "xxxxx"
-    fortran: str = "xxxxx"
-    fsharp: str = "xxxxx"
-    gedcom: str = "xxxxx"
-    gherkin: str = "xxxxx"
-    git: str = ".git"
-    glsl: str = "xxxxx"
-    go: str = ".go"
-    graphql: str = "xxxxx"
-    groovy: str = "xxxxx"
-    haml: str = "xxxxx"
-    handlebars: str = "xxxxx"
-    haskell: str = "xxxxx"
-    haxe: str = "xxxxx"
-    hpkp: str = "xxxxx"
-    hsts: str = "xxxxx"
-    http: str = "xxxxx"
-    ichigojam: str = "xxxxx"
-    icon: str = "xxxxx"
-    inform7: str = "xxxxx"
-    ini: str = ".ini"
-    io: str = "xxxxx"
-    j: str = "xxxxx"
-    java: str = ".java"
-    javascript: str = ".js"
-    jolie: str = "xxxxx"
-    json: str = ".json"
-    jsx: str = ".jsx"
-    julia: str = "xxxxx"
-    keyman: str = "xxxxx"
-    kotlin: str = ".kt"
-    latex: str = "xxxxx"
-    less: str = "xxxxx"
-    liquid: str = "xxxxx"
-    lisp: str = "xxxxx"
-    livescript: str = "xxxxx"
-    lolcode: str = "xxxxx"
-    lua: str = "xxxxx"
-    makefile: str = "xxxxx"
-    markdown: str = ".md"
-    markup_templating: str = "xxxxx"
-    markup: str = "xxxxx""xxxxx"
-    matlab: str = "xxxxx"
-    mel: str = "xxxxx"
-    mizar: str = "xxxxx"
-    monkey: str = "xxxxx"
-    n4js: str = "xxxxx"
-    nasm: str = "xxxxx"
-    nginx: str = "xxxxx"
-    nim: str = "xxxxx"
-    nix: str = "xxxxx"
-    nsis: str = "xxxxx"
-    objectivec: str = "xxxxx"
-    ocaml: str = ".ocaml"
-    opencl: str = "xxxxx"
-    oz: str = "xxxxx"
-    parigp: str = "xxxxx"
-    parser: str = "xxxxx"
-    pascal: str = "xxxxx"
-    perl: str = ".pl"
-    php_extras: str = "xxxxx"
-    php: str = ".php"
-    plsql: str = "xxxxx"
-    powershell: str = "xxxxx"
-    processing: str = "xxxxx"
-    prolog: str = "xxxxx"
-    properties: str = "xxxxx"".properties"
-    protobuf: str = "xxxxx"
-    pug: str = "xxxxx"
-    puppet: str = "xxxxx"
-    pure: str = "xxxxx"
-    python: str = ".py"
-    q: str = "xxxxx"
-    qore: str = "xxxxx"
-    r: str = "xxxxx"
-    reason: str = "xxxxx"
-    renpy: str = "xxxxx"
-    rest: str = "xxxxx"
-    rip: str = "xxxxx"
-    roboconf: str = "xxxxx"
-    ruby: str = "xxxxx"
-    rust: str = "xxxxx"
-    sas: str = "xxxxx"
-    sass: str = ".sass"
-    scala: str = ".scala"
-    scheme: str = "xxxxx"
-    scss: str = ".scss"
-    smalltalk: str = "xxxxx"
-    smarty: str = "xxxxx"
-    soy: str = "xxxxx"
-    stylus: str = "xxxxx"
-    swift: str = "xxxxx"
-    tcl: str = "xxxxx"
-    textile: str = "xxxxx"
-    tsx: str = ".tsx"
-    twig: str = "xxxxx"
-    typescript: str = "xxxxx"
-    vbnet: str = "xxxxx"
-    velocity: str = "xxxxx"
-    verilog: str = "xxxxx"
-    vhdl: str = "xxxxx"
-    vim: str = "xxxxx"
-    visual_basic: str = "xxxxx"
-    wasm: str = "xxxxx"
-    wiki: str = "xxxxx"
-    xeora: str = "xxxxx"
-    xojo: str = "xxxxx"
-    yaml: str = "xxxxx"
