+++ src/PyReprism/utils/normalizer.py
@@ -1,11 +1,9 @@
 import re
 
-class Normalizer:
+@staticmethod
+def whitespaces ():
+    return re.compile(r'[\t\x0b\x0c\r ]+')
 
-    @staticmethod
-    def whitespaces ():
-        return re.compile(r'[\t\x0b\x0c\r ]+')
-
