+++ src/tests/languages/test_python.py
@@ -1,0 +1,10 @@
+from PyReprism.languages.python import Python
+
+class TestPython:
+    str = ""
+    
+    @staticmethod
+    def test_extension():
+        ext = Python.file_extension()
+        print(ext)
+        assert(ext) == ".py"
