--- tests/languages/test_python.py
@@ -1,6 +1,0 @@
-
-
-class TestPython:
-    str = ""
-    def test_comments():
-        pass
